ai/analytics/requirements.txt
fastapi>=0.124.1,<1.0.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.0
google-cloud-bigquery==3.25.0
google-adk>=0.4.0
google-genai>=1.5.0
google-cloud-secret-manager>=2.20.0
toolbox-core>=0.3.0
matplotlib>=3.9.0
httpx>=0.27.0
